Add --mount-fd option to pass FUSE file descriptor from parent process
This adds support for a mode of operation in which a privileged parent
process opens /dev/fuse and takes care of mounting. The FUSE file
system daemon can then run as an unprivileged child that merely
processes requests on the FUSE file descriptor passed via the new
--mount-fd option.

The main benefit is that no privileged operations need to be performed
by the FUSE daemon itself directly or indirectly, so the FUSE process
can run with zero privilege. Moreover, mechanisms like securebits and
no_new_privs can be used to prevent subprocesses from re-acquiring
privilege via setuid binaries etc., reducing risk in case the FUSE
daemon gets exploited by a malicious file system.

Below is an example that illustrates this. Note that I'm using shell
for presentation purposes, the expectation is that the equivalent of
the shell commands runs in a privileged process:

  # Make binaries and libs executable by anyone
  $ chmod o+rx build/lib/libfuse3.so* example/hello
  $ export LD_LIBRARY_PATH=$PWD/build/lib

  # example/hello can mount successfully with privilege
  $ sudo sh -c "LD_LIBRARY_PATH=build/lib ./example/hello /mnt/tmp"
  $ sudo cat /mnt/tmp/hello
  Hello World!
  $ sudo umount /mnt/tmp

  # example/hello fails to mount without privilege
  $ sudo capsh --drop=all --secbits=0x3f -- -c 'LD_LIBRARY_PATH=build/lib ./example/hello -f /mnt/tmp'
  fusermount3: mount failed: Operation not permitted

  # Passing FUSE file descriptor via --mount-fd allows example/hello to work without privilege
  $ sudo sh -c '
    exec 17<>/dev/fuse
    mount -i -o nodev,nosuid,noexec,fd=17,rootmode=40000,user_id=0,group_id=0 -t fuse hello /mnt/tmp
    capsh --drop=all --secbits=0x3f -- -c "LD_LIBRARY_PATH=build/lib example/hello -f --mount-fd=17\"\""
  ' &
  [1] 55491
  $ sudo cat /mnt/tmp/hello
  Hello World!
  $ sudo umount /mnt/tmp
  [1]+  Done                    sudo sh -c '
    exec 17<>/dev/fuse
    mount -i -o nodev,nosuid,noexec,fd=17,rootmode=40000,user_id=0,group_id=0 -t fuse hello /mnt/tmp
    capsh --drop=all --secbits=0x3f -- -c "LD_LIBRARY_PATH=build/lib example/hello -f \"\""
  '